requirements.txt
asgiref==3.8.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
Django==5.2.3
django-allauth==65.11.1
django-environ==0.12.0
gunicorn==23.0.0
idna==3.10
oauthlib==3.3.1
packaging==25.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
